+++ NXNavigationExtension/UINavigationController+NXNavigationExtension.h
@@ -25,17 +25,6 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-@protocol NXNavigationExtensionInteractable <NSObject>
-
-/// 使用手势滑动返回或点击系统返回按钮过程中可以拦截或中断返回继而执行其他操作
-/// 执行 `nx_triggerSystemBackButtonHandler` 方法后也会触发以下回调
-/// @param navigationController 当前使用的导航控制器
-/// @param interactingGesture 是否为手势滑动返回
-/// @return `YES` 表示不中断返回操作继续执行；`NO` 表示中断返回操作
-- (BOOL)navigationController:(__kindof UINavigationController *)navigationController willPopViewControllerUsingInteractingGesture:(BOOL)interactingGesture;
-
-@end
-
 @interface UINavigationController (NXNavigationExtension)
 
 /// 开启全局全屏手势；默认为 NO
@@ -54,7 +43,7 @@ NS_ASSUME_NONNULL_BEGIN
 /// 执行操作之后调用 `popViewControllerAnimated:` 方法，就可以返回到指定视图控制器类型（`Class`）对应的实例中去
 /// @param aClass 指定需要跳转的视图控制器类型
 /// @param block 如果指定的视图控制器类型没有找到，则会使用回调来获取需要创建的视图控制器实例对象
-- (void)nx_redirectViewControllerClass:(Class)aClass createViewControllerUsingBlock:(__kindof UIViewController * (^__nullable)(void))block;
+- (void)nx_redirectViewControllerClass:(Class)aClass initializeStandbyViewControllerBlock:(__kindof UIViewController * (^__nullable)(void))block;
 
 @end