+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Change Log
 
+## 3.1.0
+
+- 重命名：-nx_redirectViewControllerClass:createViewControllerUsingBlock 为 -nx_redirectViewControllerClass:initializeStandbyViewControllerBlock；
+
 ## 3.0.0
 
 - 项目结构调整
 - 添加 Swift Package Manager 支持
 - 重命名：ue 前缀为 nx，防止与第三方库冲突；
-- 重命名：-navigationController:willPopViewControllerUsingInteractiveGesture 为；-navigationController:willPopViewControllerUsingInteractingGesture；
+- 重命名：-navigationController:willPopViewControllerUsingInteractiveGesture 为 -navigationController:willPopViewControllerUsingInteractingGesture；
 - 重命名：UINavigationController+UINavigationExtension 为 UINavigationController+NXNavigationExtension；
 - 重命名：UIViewController+UINavigationExtension 为 UIViewController+NXNavigationExtension；
 - 重命名：UENavigationBar 为 NXNavigationBar；